Initial commit for new libtinyev repository
Libtinyev is a tiny event's library, much in the fashion of libevent but
very very small and with less less functionality.

Most of the time, I don't need the whole funcionality that libevent provides,
but would be very nice to use epoll() in a cleanear way, by means of registering
events, function callbacks and removing this way a big loop with lots of different
things that are going on in that loop.

This library will help with that while being really small and (hopefully) well documented.

It declares only two structures, one callback function typedef and a couple handfuls of functions.